pyproject.toml
[project]
name = "ytbscript"
version = "0.1.0"
description = "YouTube video downloader service with FastAPI and yt-dlp"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "yt-dlp>=2023.11.16",
    "pydantic>=2.5.0",
    "aiohttp>=3.13.2",
    "python-dotenv>=1.2.1",
    "apscheduler>=3.10.4",
    "openai>=1.3.0",
    "requests>=2.31.0",
]
